requirements.txt
flask
python-dotenv
groq
faiss-cpu
sentence-transformers
requests
numpy
PyPDF2
